refactor(webapp): extract worker-safe toError utility [WPB-22420] (#20738)
* refactor(webapp): extract worker-safe toError utility

Move `toError` into a standalone utility module so it can be used from the
backup zip worker without importing `TypePredicateUtil`, which pulls in
app-level dependencies that are not safe in a worker context.

Update the backup worker and all `toError` consumers to import the new
utility directly, while keeping `TypePredicateUtil` focused on the
type-predicate helpers that depend on app/runtime types.

Also update tests to reflect the current `toError` behavior for plain object
errors.

Why:
- fixes the backup worker crash caused by a `window`-dependent import chain
- keeps worker code isolated from browser app dependencies
- preserves the broader app behavior while making the dependency boundary explicit

* move tests

* refactor

* fix tests
