+++ apps/webapp/src/script/auth/module/action/InvitationAction.ts
@@ -22,7 +22,7 @@ import type {NewTeamInvitation} from '@wireapp/api-client/lib/team/';
 import {Role} from '@wireapp/api-client/lib/team/member/';
 import {StatusCodes as HTTP_STATUS} from 'http-status-codes';
 
-import {toError} from 'Util/TypePredicateUtil';
+import {toError} from 'Util/toError';
 
 import {InvitationActionCreator} from './creator/';
