+++ apps/webapp/src/script/auth/module/action/LocalStorageAction.ts
@@ -17,7 +17,7 @@
  *
  */
 
-import {toError} from 'Util/TypePredicateUtil';
+import {toError} from 'Util/toError';
 
 import {LocalStorageActionCreator} from './creator/';
