+++ apps/webapp/src/script/auth/module/action/SelfAction.ts
@@ -22,7 +22,8 @@ import {StatusCodes as HTTP_STATUS} from 'http-status-codes';
 
 import {Environment} from 'Util/Environment';
 import {getLogger} from 'Util/Logger';
-import {isAxiosError, toError} from 'Util/TypePredicateUtil';
+import {toError} from 'Util/toError';
+import {isAxiosError} from 'Util/TypePredicateUtil';
 
 import {SelfActionCreator} from './creator/';
