+++ apps/webapp/src/script/auth/page/SetEmail.tsx
@@ -26,7 +26,7 @@ import {AnyAction, Dispatch} from 'redux';
 import {Button, ContainerXS, Form, H1, Input} from '@wireapp/react-ui-kit';
 
 import {t} from 'Util/LocalizerUtil';
-import {toError} from 'Util/TypePredicateUtil';
+import {toError} from 'Util/toError';
 
 import {Page} from './Page';
