+++ apps/webapp/src/script/repositories/backup/zipWorker.ts
@@ -22,7 +22,7 @@ import sodium, {ready} from 'libsodium-wrappers-sumo';
 
 import {ImportError} from './Error';
 
-import {toError} from '../../util/TypePredicateUtil';
+import {toError} from '../../util/toError';
 
 type Payload =
   | {type: 'zip'; files: Record<string, ArrayBuffer | string>; encrytionKey?: Uint8Array}
