+++ apps/webapp/src/script/repositories/event/EventService.ts
@@ -34,7 +34,7 @@ import {
 } from 'Repositories/storage';
 import {StorageSchemata} from 'Repositories/storage/StorageSchemata';
 import {getLogger, Logger} from 'Util/Logger';
-import {toError} from 'Util/TypePredicateUtil';
+import {toError} from 'Util/toError';
 
 import {ClientEvent, CONVERSATION as CLIENT_CONVERSATION_EVENT} from './Client';
